create command based on template
