+++ src/index.ts
@@ -5,16 +5,16 @@ import { Command } from "commander";
 import apply from "./commands/apply";
 import clean from "./commands/clean";
 import list from "./commands/list";
+import create from "./commands/create";
 
 const packageJson = path.join(__dirname, "..", "package.json");
 const obj = JSON.parse(fs.readFileSync(packageJson, "utf-8"));
 console.warn("K2 by github:@tuxounet");
 
 const program = new Command();
-program
-  .version(obj.version)
-  .description("An example CLI for managing a directory");
+program.version(obj.version).description("K2 Build System CLI");
 
+program.addCommand(create());
 program.addCommand(list());
 program.addCommand(apply());
 program.addCommand(clean());
